junta todos os menus no main.py e remove arquivos separados
